+++ SemiStep/Shared/Registries/IActionRegistry.cs
@@ -4,8 +4,8 @@ namespace Shared.Registries;
 
 public interface IActionRegistry
 {
-	void Initialize(IReadOnlyDictionary<short, ActionDefinition> actions);
-	ActionDefinition GetAction(short id);
+	void Initialize(IReadOnlyDictionary<int, ActionDefinition> actions);
+	ActionDefinition GetAction(int id);
 	ActionDefinition GetActionByName(string name);
 	bool ActionExists(short id);
 	bool ActionExistsByName(string name);
